+++ patzilla/access/dpma/util.py
@@ -32,6 +32,8 @@ def dpma_file_number(document_number):
     if document_number.startswith('DE'):
 
         docno_numeric = document_number[2:]
+        if not docno_numeric.isdigit():
+            raise ValueError('Document number "{}" is not numeric'.format(docno_numeric))
 
         # If document number contains a dot, assume it's
         # the checksum digit following it and we're finished.