+++ CHANGES.rst
@@ -22,8 +22,8 @@ development
     - Switch scraper from "mechanize" to "MechanicalSoup"
     - Provide command line interface program ``dpmaregister``
     - Introduce response caching with a TTL of 24 hours
-    - Decode ST.36 XML document to JSON
-    - [ui] Fix direct link to DPMA register
+    - Decode ST.36 XML documents to JSON
+    - [ui] Fix direct link to DPMAregister
 
 - Upgrade to Moment.js 2.20.0 re. CVE-2016-4055